Moved register related aarch64 translation to separate file
